fix: main-map.dart 위젯 분리 및 GetX 일부 적용
